refactored to restructure the file structure; deleted modules that are stretch goals while working towards MVP.
