+++ src/components/dialog/Dialog.css
@@ -1,0 +1,29 @@
+#addVillageForm {
+    padding: 3rem;
+}
+
+.addVillageHeader {
+    margin-top: 2rem;
+    text-align: center;
+}
+
+#buttonContainer {
+    display: flex;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+    margin-top: 1rem;
+}
+
+#closeButton {
+    position: absolute;
+    right: 1rem;
+    top: 1rem;
+}
+
+#joinVillageForm {
+    padding: 3rem;
+}
+
+#modal-header {
+    justify-content: center;
+}
