+++ src/components/village/Village.css
@@ -1,33 +1,3 @@
-#addVillageForm {
-    padding: 3rem;
-}
-
-.addVillageHeader {
-    margin-top: 2rem;
-    text-align: center;
-}
-
-#buttonContainer {
-    display: flex;
-    flex-wrap: nowrap;
-    justify-content: space-between;
-    margin-top: 1rem;
-}
-
 .calendar__event {
     border: 1px solid black;
-}
-
-#closeButton {
-    position: absolute;
-    right: 1rem;
-    top: 1rem;
-}
-
-#joinVillageForm {
-    padding: 3rem;
-}
-
-#modal-header {
-    justify-content: center;
 }
